Improved Square-Background, Added Tilesets
The Background of the Square-class can now be positioned and sized, you
can load Tilesets using the resource.manager and use them as
square-background, you can now set either the gameWidth or the
gameHeight (you get the other value as return, based on the
screen-proportions), and the orientation (wich is now fixed). Fixed some
bugs, especially regarding game-size. The Gamebook moved into the
project-folder of my test-project